+++ contracts/interfaces/ICometExtension.sol
@@ -28,7 +28,7 @@ abstract contract ICometExtension is CometCore {
         uint64 baseTrackingBorrowSpeed;
         uint104 baseMinForRewards;
         uint104 baseBorrowMin;
-        uint104 targetPercent;
+        uint64 targetPercent;
         uint104 seedReserves;
         uint104 unlockTimestamp;
         CollateralAsset[] assetConfigs;
